iot-python-mqtt/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# python-mqtt/mqttListener.py: 8
paho_mqtt == 1.6.1
